[RF] Make RooSentinel and RooMPSentinel private implementation details
The `RooSentinel` and `RooMPSentinal` classes are only helpers for the
static initialization of other RooFit classes. They should not have any
public headers, `ClassDef` macros or LinkDef entries.
