+++ roofit/roofitcore/src/RooSentinel.h
